Add JWT in the auth module #7

+++ src/modules/order/order.controller.ts
@@ -1,4 +1,16 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, HttpException, UseInterceptors } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  HttpException,
+  UseInterceptors,
+  UseGuards,
+  Req,
+} from '@nestjs/common';
 import { OrderService } from './order.service';
 import { CreateOrderDto } from './dto/create-order.dto';
 import { UpdateOrderDto } from './dto/update-order.dto';
@@ -7,7 +19,10 @@ import { UuidService } from 'src/shared/services/uuid/uuid.service';
 import { Response } from 'src/types/response';
 import { CacheInterceptor } from '@nestjs/cache-manager';
 import { PublicOrder } from './models/public-order';
+import { AuthorizationGuard } from 'src/shared/guards/authorization.guard';
+import { UserReq } from 'src/types/user-req';
 
+@UseGuards(AuthorizationGuard)
 @Controller('order')
 export class OrderController {
   constructor(
@@ -16,7 +31,8 @@ export class OrderController {
   ) {}
 
   @Post()
-  async create(@Body() createOrderDto: CreateOrderDto): Promise<Response<PublicOrder>> {
+  async create(@Req() req: UserReq, @Body() createOrderDto: CreateOrderDto): Promise<Response<PublicOrder>> {
+    const userId = req.user.sub;
     const order = new OrderEntity(
       this._uuidService.generateUuid(),
       createOrderDto.currencyCode,
@@ -24,25 +40,27 @@ export class OrderController {
       new Date(),
       new Date(),
     );
-    const possibleOrder = await this.orderService.create(order);
+    const possibleOrder = await this.orderService.create({ userId, order });
+    const orderToReturn = new PublicOrder(
+      possibleOrder.id,
+      possibleOrder.currencyCode,
+      possibleOrder.amount,
+      possibleOrder.status,
+      possibleOrder.user,
+      possibleOrder.createdAt,
+      possibleOrder.updatedAt,
+    );
     return {
       message: 'Order created successfully',
-      data: new PublicOrder(
-        possibleOrder.id,
-        possibleOrder.currencyCode,
-        possibleOrder.amount,
-        possibleOrder.status,
-        possibleOrder.user.id,
-        possibleOrder.createdAt,
-        possibleOrder.updatedAt,
-      ),
+      data: orderToReturn,
     };
   }
 
   @Get()
   @UseInterceptors(CacheInterceptor)
-  async findAll(): Promise<Response<PublicOrder[]>> {
-    const listOrders = await this.orderService.findAll();
+  async findAll(@Req() req: UserReq): Promise<Response<PublicOrder[]>> {
+    const userId = req.user.sub;
+    const listOrders = await this.orderService.findAll(userId);
     return {
       message: 'Orders retrieved successfully',
       data: listOrders.map(
@@ -52,7 +70,7 @@ export class OrderController {
             order.currencyCode,
             order.amount,
             order.status,
-            order.user.id,
+            order.user,
             order.createdAt,
             order.updatedAt,
           ),
@@ -62,8 +80,8 @@ export class OrderController {
 
   @Get(':id')
   @UseInterceptors(CacheInterceptor)
-  async findOne(@Param('id') id: string): Promise<Response<PublicOrder>> {
-    const possibleOrder = await this.orderService.findOne(id);
+  async findOne(@Req() req: UserReq, @Param('id') id: string): Promise<Response<PublicOrder>> {
+    const possibleOrder = await this.orderService.findOne({ id, userId: req.user.sub });
     if (!possibleOrder) {
       throw new HttpException('Order not found', 404);
     }
@@ -74,7 +92,7 @@ export class OrderController {
         possibleOrder.currencyCode,
         possibleOrder.amount,
         possibleOrder.status,
-        possibleOrder.user.id,
+        possibleOrder.user,
         possibleOrder.createdAt,
         possibleOrder.updatedAt,
       ),
@@ -82,8 +100,12 @@ export class OrderController {
   }
 
   @Patch(':id')
-  async update(@Param('id') id: string, @Body() updateOrderDto: UpdateOrderDto): Promise<Response<PublicOrder>> {
-    const possibleOrder = await this.orderService.update(id, updateOrderDto);
+  async update(
+    @Req() req: UserReq,
+    @Param('id') id: string,
+    @Body() updateOrderDto: UpdateOrderDto,
+  ): Promise<Response<PublicOrder>> {
+    const possibleOrder = await this.orderService.update({ userId: req.user.sub, id, updateOrderDto });
     if (!possibleOrder) {
       throw new HttpException('Order not found', 404);
     }
@@ -94,7 +116,7 @@ export class OrderController {
         possibleOrder.currencyCode,
         possibleOrder.amount,
         possibleOrder.status,
-        possibleOrder.user.id,
+        possibleOrder.user,
         possibleOrder.createdAt,
         possibleOrder.updatedAt,
       ),
@@ -102,8 +124,8 @@ export class OrderController {
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string): Promise<Response<boolean>> {
-    const status = await this.orderService.remove(id);
+  async remove(@Req() req: UserReq, @Param('id') id: string): Promise<Response<boolean>> {
+    const status = await this.orderService.remove({ userId: req.user.sub, id });
     if (!status) {
       throw new HttpException('Order not found', 404);
     }
